changestreams/requirements.txt
dnspython==2.1.0
pymongo==3.12.1
python-dotenv==0.20.0
